implemented customers CRUD

+++ routes/api.php
@@ -4,6 +4,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\UserController;
+use App\Http\Controllers\CustomerController;
 
 /*
 |--------------------------------------------------------------------------
@@ -25,5 +26,17 @@ Route::middleware(['auth:sanctum'])->group(function() {
   Route::get('/users', [UserController::class, 'index']);
   Route::get('/users/{id}', [UserController::class, 'show']);
   Route::put('/users/{id}', [UserController::class, 'update']);
-  Route::delete('/users/{id}', [UserController::class, 'destroy']);
+
+  Route::get('/customers', [CustomerController::class, 'index']);
+  Route::post('/customers', [CustomerController::class, 'store']);
+  Route::get('/customers/{id}', [CustomerController::class, 'show']);
+  Route::put('/customers/{id}', [CustomerController::class, 'update']);
+
+  // User admin routes
+  Route::middleware(['user_is_admin'])->group(function() {
+    Route::delete('/admin/users/{id}', [UserController::class, 'destroy']);
+    Route::put('/admin/users/{id}', [UserController::class, 'toggleAdminAccess']);
+
+    Route::delete('/admin/customers/{id}', [CustomerController::class, 'destroy']);
+  });
 });
